requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
azure-common==1.1.28
azure-core==1.30.2
azure-core-tracing-opentelemetry==1.0.0b11
azure-cosmos==4.7.0
azure-identity==1.17.1
azure-keyvault-secrets==4.8.0
azure-monitor-opentelemetry==1.6.0
azure-monitor-opentelemetry-exporter==1.0.0b27
azure-search==1.0.0b2
azure-search-documents==11.5.2
azure-storage-blob==12.20.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coverage==7.6.0
cryptography==42.0.8
dataclasses-json==0.6.7
Deprecated==1.2.14
distlib==0.3.8
distro==1.9.0
filelock==3.15.1
fixedint==0.1.6
Flask==3.0.3
frozenlist==1.4.1
google-api-core==2.19.2
google-auth==2.34.0
googleapis-common-protos==1.65.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
idna==3.7
importlib_metadata==7.1.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
jsonify==0.5
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.37
langchain-openai==0.3.6
langchain-text-splitters==0.3.6
langgraph==0.2.74
langgraph-checkpoint==2.0.16
langgraph-sdk==0.1.53
langsmith==0.3.8
MarkupSafe==2.1.5
marshmallow==3.26.1
msal==1.29.0
msal-extensions==1.2.0
msgpack==1.1.0
msrest==0.7.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.63.2
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.13
opentelemetry-api==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-dbapi==0.46b0
opentelemetry-instrumentation-django==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-instrumentation-flask==0.46b0
opentelemetry-instrumentation-psycopg2==0.46b0
opentelemetry-instrumentation-requests==0.46b0
opentelemetry-instrumentation-urllib==0.46b0
opentelemetry-instrumentation-urllib3==0.46b0
opentelemetry-instrumentation-wsgi==0.46b0
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
orjson==3.10.15
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
portalocker==2.10.0
proto-plus==1.24.0
protobuf==5.28.1
psutil==5.9.8
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.7.4
pydantic-settings==2.7.1
pydantic_core==2.18.4
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.5.15
requests==2.32.3
requests-mock==1.12.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retrying==1.3.4
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.38
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
virtualenv==20.26.2
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2
zstandard==0.23.0
uvicorn